Use individual outcomes for R111 and R113
