+++ packages/alfa-rules/test/sia-r111/rule.spec.tsx
@@ -3,11 +3,10 @@ import { test } from "@siteimprove/alfa-test";
 
 import { Device } from "@siteimprove/alfa-device";
 
-import R111 from "../../src/sia-r111/rule";
+import R111, { Outcomes } from "../../src/sia-r111/rule";
 
 import { evaluate } from "../common/evaluate";
 import { passed, failed, inapplicable } from "../common/outcome";
-import { BoundingBox } from "../../src/common/outcome/bounding-box";
 
 test("evaluate() passes button with clickable area of exactly 44x44 pixels", async (t) => {
   const device = Device.standard();
@@ -25,7 +24,7 @@ test("evaluate() passes button with clickable area of exactly 44x44 pixels", asy
 
   t.deepEqual(await evaluate(R111, { document, device }), [
     passed(R111, target, {
-      1: BoundingBox.HasSufficientSize(
+      1: Outcomes.HasSufficientSize(
         "Hello",
         target.getBoundingBox(device).getUnsafe(),
       ),
@@ -47,7 +46,7 @@ test("evaluate() passes input element regardless of size", async (t) => {
 
   t.deepEqual(await evaluate(R111, { document, device }), [
     passed(R111, target, {
-      1: BoundingBox.IsUserAgentControlled("", target.getBoundingBox(device).getUnsafe()),
+      1: Outcomes.IsUserAgentControlled("", target.getBoundingBox(device).getUnsafe()),
     }),
   ]);
 });
@@ -68,7 +67,7 @@ test("evaluate() fails button with clickable area of less than 44x44 pixels", as
 
   t.deepEqual(await evaluate(R111, { document, device }), [
     failed(R111, target, {
-      1: BoundingBox.HasInsufficientSize(
+      1: Outcomes.HasInsufficientSize(
         "Hello",
         target.getBoundingBox(device).getUnsafe(),
       ),
